requirements.txt
torch
torchvision
timm
numpy
pandas
openpyxl
scikit-learn
opencv-python
Pillow
matplotlib
tqdm
python-chess
cairosvg